+++ src/lib/waku_message/topic_only_message.ts
@@ -6,6 +6,10 @@ import type { Decoder, Message, ProtoMessage } from "../interfaces";
 const log = debug("waku:message:topic-only");
 
 export class TopicOnlyMessage implements Message {
+  public payload: undefined;
+  public rateLimitProof: undefined;
+  public timestamp: undefined;
+
   constructor(private proto: proto.TopicOnlyMessage) {}
 
   get contentTopic(): string {
@@ -19,7 +23,13 @@ export class TopicOnlyDecoder implements Decoder<TopicOnlyMessage> {
   decodeProto(bytes: Uint8Array): Promise<ProtoMessage | undefined> {
     const protoMessage = proto.TopicOnlyMessage.decode(bytes);
     log("Message decoded", protoMessage);
-    return Promise.resolve(protoMessage);
+    return Promise.resolve({
+      contentTopic: protoMessage.contentTopic,
+      payload: undefined,
+      rateLimitProof: undefined,
+      timestamp: undefined,
+      version: undefined,
+    });
   }
 
   async decode(proto: ProtoMessage): Promise<TopicOnlyMessage | undefined> {